refactor(app): run initial prettier + eslint

+++ code13/app/client/imports/components/Assets/EditActor/Forms/Animations.js
@@ -6,76 +6,73 @@ import SmallDD from '../../../Controls/SmallDD.js'
 import MgbActor from '/client/imports/components/MapActorGameEngine/MageMgbActor'
 
 export default class Animations extends React.Component {
-
   state = {
     serializedForm: {},
     isLoading: false, // When loading graphic asset tileset info and frames
     showModal: false, // When importing graphic frames
     graphicFrameImports: [], // Data of frames to import; contains name and if frame is selected
     animationIndex: 0, // Starting animation frame being changed
-    lastEffectChange: ""
+    lastEffectChange: '',
   }
 
   get data() {
     return this.props.asset.content2.animationTable
   }
 
-  handleChange = (e, { value }) => this.setState({ value })  // TODO: How is this called?
+  handleChange = (e, { value }) => this.setState({ value }) // TODO: How is this called?
 
-  handleSubmit = (e, serializedForm) => {   // TODO: How is this called?
+  handleSubmit = (e, serializedForm) => {
+    // TODO: How is this called?
     e.preventDefault()
     this.setState({ serializedForm })
   }
 
   handleGraphicFrameSelection(e, data) {
     let frames = this.state.graphicFrameImports
-    if (!data.checked)
-      frames[data.value].checked = 0
-    else
-      frames[data.value].checked = 1
+    if (!data.checked) frames[data.value].checked = 0
+    else frames[data.value].checked = 1
 
-    this.setState({graphicFrameImports: frames})
+    this.setState({ graphicFrameImports: frames })
   }
 
   handleClearAll(animTitle) {
     _.forEach(Object.keys(this.refs), frame => {
-      if (_.startsWith(frame, animTitle))
-        this.refs[frame].handleRemove()
+      if (_.startsWith(frame, animTitle)) this.refs[frame].handleRemove()
     })
   }
 
   handleApplyEffects(index, animTitle) {
-    if (!this.state.lastEffectChange)
-      return
+    if (!this.state.lastEffectChange) return
 
     let count = 0
-    if (_.startsWith(animTitle, "move"))
-      count = 5
-    else if (_.startsWith(animTitle, "melee"))
-      count = 8
-    else if (_.startsWith(animTitle, "stationary"))
-      count = 16
+    if (_.startsWith(animTitle, 'move')) count = 5
+    else if (_.startsWith(animTitle, 'melee')) count = 8
+    else if (_.startsWith(animTitle, 'stationary')) count = 16
 
     // Get indices of animation frames in the animation group
     // index is the index of the animation frame in the data array of all animation frames
     // i is the index of the last frame for the animation group animTitle
     // frameIndex is the index of the current frame relative to the animaton group
-    let frameIndex = _.parseInt(this.data[index].action.split(" ").pop())
+    let frameIndex = _.parseInt(this.data[index].action.split(' ').pop())
 
-    if (_.startsWith(this.data[index].action, "step")) // face + step 1, 2, 3, 4
+    if (
+      _.startsWith(this.data[index].action, 'step') // face + step 1, 2, 3, 4
+    )
       frameIndex++
-    else if (!_.isInteger(frameIndex)) // All animations except 'face ___' have a frame number in action
+    else if (
+      !_.isInteger(frameIndex) // All animations except 'face ___' have a frame number in action
+    )
       frameIndex = 1
 
-    for (let i=1; i<=count; i++)
-      this.data[i + index - frameIndex].effect = this.state.lastEffectChange
+    for (let i = 1; i <= count; i++) this.data[i + index - frameIndex].effect = this.state.lastEffectChange
 
     this.forceUpdate()
     this.props.onChange && this.props.onChange()
   }
 
-  onChange = (e) => {   // TODO: How is this called?
-    console.log("Form has changed")
+  onChange = e => {
+    // TODO: How is this called?
+    console.log('Form has changed')
   }
 
   /*
@@ -88,50 +85,47 @@ export default class Animations extends React.Component {
   // Store graphic frames in graphicFrameImports
   getGraphicFrames(data, val) {
     let graphicFrames = []
-    for (let i=0; i<data.tilecount; i++) {
-      let name = val + ' #' + (i+1)
-      let frame = {name: name, checked: 1}
+    for (let i = 0; i < data.tilecount; i++) {
+      let name = val + ' #' + (i + 1)
+      let frame = { name: name, checked: 1 }
       graphicFrames.push(frame)
     }
-    this.setState({graphicFrameImports: graphicFrames})
+    this.setState({ graphicFrameImports: graphicFrames })
   }
 
   // Change graphics based on selected frames from graphicFrameImports
   changeGraphicFrames() {
     let index = this.state.animationIndex
     let frames = this.state.graphicFrameImports
-    for (let i=0; i<frames.length; i++) {
+    for (let i = 0; i < frames.length; i++) {
       if (frames[i].checked) {
-        this.data[index+i] = {
-          "action": MgbActor.animationNames[index+i],
-          "tileName": frames[i].name,
-          "frame": i,
-          "effect": "no effect"
+        this.data[index + i] = {
+          action: MgbActor.animationNames[index + i],
+          tileName: frames[i].name,
+          frame: i,
+          effect: 'no effect',
         }
-      }
-      else
-        index--
+      } else index--
     }
-    this.setState({showModal: false, graphicFrameImports: [], animationIndex: 0})
+    this.setState({ showModal: false, graphicFrameImports: [], animationIndex: 0 })
     this.props.onChange && this.props.onChange()
   }
 
   changeGraphic(index, val, asset) {
     this.data[index].tileName = val
     this.data[index].frame = 0
-    this.setState({animationIndex: index})
+    this.setState({ animationIndex: index })
 
     if (asset) {
-      this.setState({isLoading: true})
+      this.setState({ isLoading: true })
 
-      $.get('/api/asset/tileset-info/' + asset._id, (data) => {
+      $.get('/api/asset/tileset-info/' + asset._id, data => {
         if (data.tilecount > 1) {
-          this.setState({showModal: true})
+          this.setState({ showModal: true })
           this.getGraphicFrames(data, val)
         }
-      })
-      .done(() => {
-        this.setState({isLoading: false})
+      }).done(() => {
+        this.setState({ isLoading: false })
       })
     }
 
@@ -139,7 +133,7 @@ export default class Animations extends React.Component {
   }
 
   changeEffect(index, val) {
-    this.setState({lastEffectChange: val})
+    this.setState({ lastEffectChange: val })
 
     this.data[index].effect = val
 
@@ -149,8 +143,10 @@ export default class Animations extends React.Component {
 
   renderContent(animations, animTitle, i) {
     return (
-      <Table.Row style={!this.props.canEdit ? {pointerEvents: 'none'} : {}} key={i}>
-        <Table.Cell>{animations[i]}</Table.Cell>
+      <Table.Row style={!this.props.canEdit ? { pointerEvents: 'none' } : {}} key={i}>
+        <Table.Cell>
+          {animations[i]}
+        </Table.Cell>
         <Table.Cell>
           <DropArea
             kind="graphic"
@@ -160,28 +156,28 @@ export default class Animations extends React.Component {
             asset={this.props.asset}
             isLoading={this.state.isLoading}
             onChange={this.changeGraphic.bind(this, i)}
-            ref={animTitle + " " + i}
+            ref={animTitle + ' ' + i}
           />
         </Table.Cell>
-        <Table.Cell style={{overflow: 'visible'}}>
-          <SmallDD  options={MgbActor.animationEffectNames} value={this.data[i].effect} onChange={this.changeEffect.bind(this, i)} />
+        <Table.Cell style={{ overflow: 'visible' }}>
+          <SmallDD
+            options={MgbActor.animationEffectNames}
+            value={this.data[i].effect}
+            onChange={this.changeEffect.bind(this, i)}
+          />
         </Table.Cell>
       </Table.Row>
     )
   }
 
   renderAccordion(animTable, animTitle, i) {
-    if (animTitle !== "stationary west")
-      i--
+    if (animTitle !== 'stationary west') i--
 
     return (
       <div key={i}>
-        <Accordion styled fluid
-          exclusive={animTitle === "stationary"}
-          defaultActiveIndex={0}
-        >
+        <Accordion styled fluid exclusive={animTitle === 'stationary'} defaultActiveIndex={0}>
           <Accordion.Title>
-            <Icon name='dropdown' />
+            <Icon name="dropdown" />
             {animTitle}
           </Accordion.Title>
           <Accordion.Content>
@@ -191,28 +187,37 @@ export default class Animations extends React.Component {
                   <Table.HeaderCell>Animation Frame</Table.HeaderCell>
                   <Table.HeaderCell width={8}>
                     Graphic
-                    <Button compact
+                    <Button
+                      compact
                       size="mini"
-                      floated='right'
+                      floated="right"
                       onClick={() => this.handleClearAll(animTitle)}
                     >
                       Clear All
-                  </Button>
+                    </Button>
                   </Table.HeaderCell>
-                  <Table.HeaderCell style={{position:'relative'}}>
+                  <Table.HeaderCell style={{ position: 'relative' }}>
                     Orientation
-                    <Button compact
+                    <Button
+                      compact
                       size="mini"
-                      floated='right'
+                      floated="right"
                       onClick={() => this.handleApplyEffects(i, animTitle)}
                     >
                       Apply to All
                     </Button>
                     {this.state.lastEffectChange &&
-                      <span style={{position:'absolute',bottom:'-3px',right:'10px',fontSize:'8px',color:'grey'}}>
+                      <span
+                        style={{
+                          position: 'absolute',
+                          bottom: '-3px',
+                          right: '10px',
+                          fontSize: '8px',
+                          color: 'grey',
+                        }}
+                      >
                         Effect: {this.state.lastEffectChange}
-                      </span>
-                    }
+                      </span>}
                   </Table.HeaderCell>
                 </Table.Row>
               </Table.Header>
@@ -236,83 +241,93 @@ export default class Animations extends React.Component {
     let animTable = []
     let prevDirection = animations[0].split(' ')[1]
     let name = animations[0]
-    let animTitle = _.startsWith(name, 'face') ? 'move ' + name.split(' ')[1] : (name.split(' ').length > 2 ? name.split(' ')[0] + ' ' + name.split(' ')[1] : name.split(' ')[0])
+    let animTitle = _.startsWith(name, 'face')
+      ? 'move ' + name.split(' ')[1]
+      : name.split(' ').length > 2 ? name.split(' ')[0] + ' ' + name.split(' ')[1] : name.split(' ')[0]
 
-    for (let i=0; i<animations.length; i++) {
+    for (let i = 0; i < animations.length; i++) {
       name = animations[i]
       if (!this.data[i]) {
         this.data[i] = {
-          "action": name,
-          "tileName": null,
-          "frame": 0,
-          "effect": "no effect",
+          action: name,
+          tileName: null,
+          frame: 0,
+          effect: 'no effect',
         }
       }
 
       if (
-        !((aType === '0' || aType === '1') && (_.startsWith(name, 'stationary') && name.split(' ').length === 2)) && // Filter out stationary for Player/NPC
+        !(
+          (aType === '0' || aType === '1') &&
+          (_.startsWith(name, 'stationary') && name.split(' ').length === 2)
+        ) && // Filter out stationary for Player/NPC
         !(aType === '3' && (_.startsWith(name, 'stationary') || _.startsWith(name, 'melee'))) && // Filter out non-movement for Shot
-        !(['2', '4', '5', '6', '7'].indexOf(aType) > -1 && (!_.startsWith(name, 'stationary') || name.split(' ').length !== 2)) // Filter out non-stationary for Item/Wall/Floor/Scenery
+        !(
+          ['2', '4', '5', '6', '7'].indexOf(aType) > -1 &&
+          (!_.startsWith(name, 'stationary') || name.split(' ').length !== 2)
+        ) // Filter out non-stationary for Item/Wall/Floor/Scenery
       ) {
         // Group animations by direction
         if (
           (['2', '4', '5', '6', '7'].indexOf(aType) === -1 && name.includes(prevDirection)) ||
-          (['2', '4', '5', '6', '7'].indexOf(aType) > -1  && _.startsWith(name, 'stationary') && name.split(' ').length === 2) || // Don't use prevDirection if only stationary animations
-          i+1 === animations.length
+          (['2', '4', '5', '6', '7'].indexOf(aType) > -1 &&
+            _.startsWith(name, 'stationary') &&
+            name.split(' ').length === 2) || // Don't use prevDirection if only stationary animations
+          i + 1 === animations.length
         ) {
-          if (['2', '4', '5', '6', '7'].indexOf(aType) !== -1)
-            animTitle = "stationary"
+          if (['2', '4', '5', '6', '7'].indexOf(aType) !== -1) animTitle = 'stationary'
 
           animTable.push(this.renderContent(animations, animTitle, i))
           // Fencepost
-          if (i+1 === animations.length)
-            rows.push(this.renderAccordion(animTable, animTitle, i))
-        }
-        // Put animation group for current direction in accordion
-        else
-        {
-          if (animTable.length > 0)
-            rows.push(this.renderAccordion(animTable, animTitle, i))
+          if (i + 1 === animations.length) rows.push(this.renderAccordion(animTable, animTitle, i))
+        } else {
+          // Put animation group for current direction in accordion
+          if (animTable.length > 0) rows.push(this.renderAccordion(animTable, animTitle, i))
 
           // Content for next direction
           prevDirection = name.split(' ')[1] // get direction from animation name which is the 2nd part of string (action direction frameNum)
-          animTitle = _.startsWith(name, 'face') ? 'move ' + name.split(' ')[1] : (name.split(' ').length > 2 ? name.split(' ')[0] + ' ' + name.split(' ')[1] : name.split(' ')[0])
+          animTitle = _.startsWith(name, 'face')
+            ? 'move ' + name.split(' ')[1]
+            : name.split(' ').length > 2 ? name.split(' ')[0] + ' ' + name.split(' ')[1] : name.split(' ')[0]
           animTable = []
 
           animTable.push(this.renderContent(animations, animTitle, i))
         }
-      }
-      else if (animTable.length > 0) {
+      } else if (animTable.length > 0) {
         if (['2', '4', '5', '6', '7'].indexOf(aType) > -1)
           rows.push(this.renderAccordion(animTable, 'stationary', i))
-        else
-          rows.push(this.renderAccordion(animTable, animTitle, i))
+        else rows.push(this.renderAccordion(animTable, animTitle, i))
 
         animTable = []
       }
     }
 
     return (
-      <div style={{position: 'relative'}}>
-        {
-          this.state.isLoading &&
+      <div style={{ position: 'relative' }}>
+        {this.state.isLoading &&
           <Dimmer active inverted>
-            <Loader style={{position: 'fixed', right: '345px', top: '50%', translate: "transform(-50%, -50%)"}} active inline size='large'>Loading frames...</Loader>
-          </Dimmer>
-        }
-        {
-          this.state.showModal &&
-          <Modal defaultOpen
+            <Loader
+              style={{ position: 'fixed', right: '345px', top: '50%', translate: 'transform(-50%, -50%)' }}
+              active
+              inline
+              size="large"
+            >
+              Loading frames...
+            </Loader>
+          </Dimmer>}
+        {this.state.showModal &&
+          <Modal
+            defaultOpen
             size="small"
-            onUnmount={() => {this.setState({showModal: false, graphicFrameImports: [], animationIndex: 0})}}
+            onUnmount={() => {
+              this.setState({ showModal: false, graphicFrameImports: [], animationIndex: 0 })
+            }}
           >
             <Modal.Header>Import Graphic Frames</Modal.Header>
             <Modal.Content>
               <Item.Group divided>
-                {
-                _.map(this.state.graphicFrameImports, (frame, i) => {
-                  if (!frame.checked)
-                    checkedCount--
+                {_.map(this.state.graphicFrameImports, (frame, i) => {
+                  if (!frame.checked) checkedCount--
 
                   return (
                     <Item key={i}>
@@ -321,37 +336,32 @@ export default class Animations extends React.Component {
                         value={frame.name}
                         frame={i}
                         asset={this.props.asset}
-                        isModalView={true}
+                        isModalView
                       />
                       <Checkbox
                         name={frame.name}
                         value={i}
-                        style={{position: 'absolute', right: '25px'}}
+                        style={{ position: 'absolute', right: '25px' }}
                         defaultChecked
-                        onChange={(e, data) => {this.handleGraphicFrameSelection(e, data)}}
+                        onChange={(e, data) => {
+                          this.handleGraphicFrameSelection(e, data)
+                        }}
                       />
                     </Item>
-                  )}
                   )
-                }
+                })}
               </Item.Group>
             </Modal.Content>
             <Modal.Actions>
-              <Button primary
-                disabled={!checkedCount}
-                onClick={() => this.changeGraphicFrames()}>
+              <Button primary disabled={!checkedCount} onClick={() => this.changeGraphicFrames()}>
                 Import Selected Frames
               </Button>
             </Modal.Actions>
-          </Modal>
-        }
-        {
-          rows.map((anim) => {
-            return anim
-          })
-        }
+          </Modal>}
+        {rows.map(anim => {
+          return anim
+        })}
       </div>
     )
   }
 }
-
